Переносит всю логику по отрисовке основного содержимого страницы в контроллер pageController
