+++ src/main.js
@@ -1,107 +1,18 @@
-import CardComponent from "./components/card.js";
-import CardsListComponent from "./components/cards-list.js";
-import FilmDetailsComponent from "./components/film-details.js";
 import FilterComponent from "./components/filter.js";
 import FooterStatComponent from "./components/footer-stat.js";
-import MostCommentedComponent from "./components/most-commented.js";
-import NoFilmsComponent from "./components/no-films.js";
-import ShowMoreComponent from "./components/show-more.js";
+import PageController from "./controllers/page-controller.js";
 import SiteNavComponent from "./components/site-nav.js";
-import SortComponent from "./components/sort.js";
-import TopRateComponent from "./components/top-rate.js";
 import UserProfileComponent from "./components/user-profile.js";
 import {generateFilms} from "./mock/film.js";
-import {getUserLevel, getExtraFilms} from "./utils/components-data.js";
-import {render, remove} from "./utils/render.js";
+import {getUserLevel} from "./utils/components-data.js";
+import {render} from "./utils/render.js";
 
-const SHOWING_CARDS_AMOUNT_ON_START = 5;
-const SHOWING_CARDS_AMOUNT_BY_BUTTON = 5;
-const CARDS_AMOUNT_EXTRA = 2;
 const FILMS_AMOUNT = 20;
 const films = generateFilms(FILMS_AMOUNT);
 const userLevel = getUserLevel(films);
 const bodyElement = document.querySelector(`body`);
 const pageHeaderElement = bodyElement.querySelector(`.header`);
 const pageMainElement = bodyElement.querySelector(`.main`);
-
-const renderCard = (cardsListContainer, film) => {
-  const cardComponent = new CardComponent(film);
-  const filmDetailsComponent = new FilmDetailsComponent(film);
-
-  const openPopup = () => {
-    render(bodyElement, filmDetailsComponent);
-    filmDetailsComponent.setCloseButtonClickHandler(onCloseButtonClick);
-    document.addEventListener(`keydown`, onEscPress);
-  };
-
-  const closePopup = () => {
-    remove(filmDetailsComponent);
-  };
-
-  const onCloseButtonClick = () => {
-    closePopup();
-    document.removeEventListener(`keydown`, onEscPress);
-  };
-
-  const onEscPress = (evt) => {
-    evt.preventDefault();
-    if (evt.key === `Escape`) {
-      closePopup();
-      document.removeEventListener(`keydown`, onEscPress);
-    }
-  };
-
-  render(cardsListContainer, cardComponent);
-  cardComponent.setClickHandlers(() => openPopup());
-};
-
-const renderMainBoard = () => {
-  // Отрисовываю сортировку
-  render(pageMainElement, new SortComponent());
-
-  // Сообщение об отсутствии фильмов в системе, если их нет
-  if (!films.length) {
-    render(pageMainElement, new NoFilmsComponent());
-    return;
-  }
-  //  Контейнер для карточек фильмов
-  const cardsListComponent = new CardsListComponent();
-  render(pageMainElement, cardsListComponent);
-  const listContainer = cardsListComponent.getElement().querySelector(`.films-list__container`);
-  let showingCardsCount = SHOWING_CARDS_AMOUNT_ON_START;
-  for (let i = 0; i < showingCardsCount; i++) {
-    renderCard(listContainer, films[i]);
-  }
-
-  // Кнопка для открытия следующей порции карточек
-  const filmsListElement = cardsListComponent.getElement().querySelector(`.films-list`);
-  const showMoreComponent = new ShowMoreComponent();
-  render(filmsListElement, showMoreComponent);
-
-  showMoreComponent.getElement().addEventListener(`click`, () => {
-    const previousCardsCount = showingCardsCount;
-    showingCardsCount += SHOWING_CARDS_AMOUNT_BY_BUTTON;
-    films.slice(previousCardsCount, showingCardsCount)
-    .forEach((film) => renderCard(listContainer, film));
-    if (showingCardsCount >= FILMS_AMOUNT) {
-      remove(showMoreComponent);
-    }
-  });
-
-  // Дополнительные секции
-  render(cardsListComponent.getElement(), new TopRateComponent());
-  render(cardsListComponent.getElement(), new MostCommentedComponent());
-
-  const filmListExtraElements = cardsListComponent.getElement().querySelectorAll(`.films-list--extra`);
-  const extraFilms = getExtraFilms(films, CARDS_AMOUNT_EXTRA);
-  let count = 0;
-  filmListExtraElements.forEach((listElement) => {
-    const container = listElement.querySelector(`.films-list__container`);
-    extraFilms[count].forEach((film) => renderCard(container, film));
-    count++;
-  });
-
-};
 
 render(pageHeaderElement, new UserProfileComponent(userLevel));
 
@@ -110,7 +21,9 @@ const siteNavComponent = new SiteNavComponent();
 render(pageMainElement, siteNavComponent);
 render(siteNavComponent.getElement(), new FilterComponent(films), `afterbegin`);
 
-renderMainBoard();
+// Отрисовываю основное содержимое страницы
+const pageController = new PageController(pageMainElement, bodyElement);
+pageController.render(films);
 
 // Статистика в футере
 const footerStatisticsElement = bodyElement.querySelector(`.footer__statistics`);
